backend/requirements.txt
Django>=4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.6,<3.0
drf-spectacular>=0.26.2,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3